SA-010: Manage Siswa Data & Show Nilai (#22)
* feat: implement Siswa management with CRUD operations

- Added SiswaController for handling student-related actions.
- Created CreateSiswaRequest and EditSiswaRequest for validation.
- Developed SiswaService for business logic related to students.
- Implemented views for listing, creating, editing, and showing student details.
- Updated navigation and routes for Siswa management.

* fix: correct status transition logic in SiswaService

- Updated the status transition for 'block' to change to 'active' instead of 'pending' in the toggleActive method.

* feat: add RapotHafalanController and related services for managing student hafalan reports

- Introduced `RapotHafalanController` to handle CRUD operations for student hafalan reports.
- Created `RapotHafalanService` to encapsulate business logic related to hafalan reports.
- Added `RaportService` and `RaportNilaiService` for managing raport and nilai data.
- Updated `SiswaController` to integrate new services for enhanced student detail views.
- Enhanced the `show` view for students to display detailed raport and nilai information.

This commit lays the groundwork for comprehensive management of student hafalan reports and integrates it with existing student functionalities.

* feat(raport): add student ranking and average score calculation

This commit introduces new methods to calculate the average score and ranking of students based on their academic performance. The changes include adding `averageRaportNilaiByRaportId` and `getRankingRaportNilaiByRaportId` methods in `RaportNilaiService`, and updating the `show` method in `SiswaController` to fetch and display these metrics. The view has also been updated to reflect the new data.

* feat(siswa): enhance SiswaController and related services for raport management

This commit introduces the `RaportSikapService` to manage student behavior reports and integrates it into the `SiswaController`. The `show` method is updated to fetch and display behavior reports alongside academic performance metrics. Additionally, the `RaportNilaiService` methods are refined for better clarity, and the view is adjusted to present the new data structure effectively. This enhancement improves the overall student report management experience.

* feat(siswa): add export functionality for student reports in CSV and Excel formats

This commit introduces two new methods in the `SiswaController` for exporting student reports: `exportCsv` and `exportExcel`. The `ExportService` is created to handle the logic for generating CSV and Excel files from the raport data. Additionally, the routes for these export functionalities are defined in `admin.php`, and the student detail view is updated to include buttons for exporting reports. This enhancement improves the reporting capabilities for student data management.

* fix: fixed ui and form

---------

Co-authored-by: Miftahul Huda <iniakunhuda@gmail.com>